revert ad45f0e, which breaks :CommandT with a path argument
